requirements.txt
#
# This file is autogenerated by pip-compile with python 3.10
# To update, run:
#
#    pip-compile requirements.in
#
asgiref==3.5.2
    # via django
cffi==1.15.1
    # via cryptography
cryptography==38.0.3
    # via djangorestframework-simplejwt
defusedxml==0.7.1
    # via djangorestframework-xml
django==4.1.3
    # via
    #   -r requirements.in
    #   django-filter
    #   djangorestframework
    #   djangorestframework-simplejwt
django-filter==22.1
    # via -r requirements.in
djangorestframework==3.14.0
    # via
    #   -r requirements.in
    #   djangorestframework-simplejwt
djangorestframework-simplejwt[crypto]==5.2.2
    # via -r requirements.in
djangorestframework-xml==2.0.0
    # via -r requirements.in
pillow==9.3.0
    # via -r requirements.in
pycparser==2.21
    # via cffi
pyjwt==2.6.0
    # via djangorestframework-simplejwt
pytz==2022.6
    # via djangorestframework
sqlparse==0.4.3
    # via django
